Update API routes.
